1. Замержил код; 2. Добавил в ClientBattleMarketManager методы browseRandomBattleMap и browseBattleMaps; 3. Переименовал и доработал классы ClientBrowseBattleMap*Answer; 4. Доделал тесты battleConnectionTest; 5. Поправил баг, когда клиент запрашивал одну случайную карту; 6. Временно переименовал файл ClientBrowseBAttleMapsRequest на ClientBrowseBattleMapsRequest1.

+++ client/src/com/geargames/regolith/serializers/BattleMapDeserializer.java
@@ -149,10 +149,10 @@ public class BattleMapDeserializer {
     }
 
 
-    public static BattleMap  deserializeLightBattleMap(MicroByteBuffer buffer, BaseConfiguration baseConfiguration){
+    public static BattleMap deserializeLightBattleMap(MicroByteBuffer buffer, BaseConfiguration baseConfiguration){
         BattleMap battleMap = new BattleMap();
         battleMap.setId(SimpleDeserializer.deserializeInt(buffer));
-        int amount = buffer.get();
+        byte amount = buffer.get();
         BattleType[] battleTypes = new BattleType[amount];
         battleMap.setPossibleBattleTypes(battleTypes);
         for (int j = 0; j < amount; j++) {
